OpenGL tutorials solution: creating shader in tutorial project (stage #17);
(1) Rendering scene class is added;
(2) Shader code source interface is improved;
The scene still be in development stage yet;

+++ project.cpp.tutors/open_gl/open_gl.shared/interface/shader/gl_shd_type.h
@@ -41,7 +41,7 @@ namespace ex_ui { namespace draw { namespace open_gl { namespace shader {
 
 		operator uint16_t (void) const; // returns currently set value of the shader type;
 
-		static CString To_str(const uint16_t _u_type);
+		static CString To_str(const uint32_t _u_type); // the data type is intentionally changed for avoiding unnecessary casts;
 		static e_value Get_type (const uint32_t _u_shader_id, CError&);
 
 	public:
